Potentially horrible solution for type adapter privacy.
