+++ src/main/java/qupath/bioimageio/spec/tensor/sizes/Sizes.java
@@ -1,0 +1,4 @@
+package qupath.bioimageio.spec.tensor.sizes;
+
+public class Sizes {
+}
